+++ crates/weaver_resolver/src/error.rs
@@ -184,7 +184,7 @@ pub enum Error {
     },
 
     /// A container for multiple errors.
-    #[error("{:?}", format_errors(.0))]
+    #[error("{}", format_errors(.0))]
     CompoundError(#[related] Vec<Error>),
 }